Extract memberList and memberlistitem
